requirements.txt
# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data processing
pandas==2.1.4
numpy==1.26.2

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1